Camera Go Zoom, also killing last Player
Camera zooms in and out depending on where the players are, can be tweaked in the CameraDistance Curve. Also added functionality to destroy the last player if they fall behind too far (distance can be set on Camera). The kill logic needs to be moved to an appropriate space, currently in Camera. Also added a Blueprint Library for custom functions